requirements.txt
# Core dependencies
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
numpy>=1.21.0
matplotlib>=3.3.0
Pillow>=8.0.0
scikit-image>=0.18.0
scikit-learn>=1.0.0
transformers>=4.20.0
tqdm>=4.60.0
pandas>=1.3.0
seaborn>=0.11.0

# SAM2 for spheroid segmentation (required)
segment-anything-2